Rename package to start with lower case
