+++ androidApp/app/src/androidTest/java/io/github/nadzwyczajnaGrupaRobocza/texaspoker/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package io.github.NadzwyczajnaGrupaRobocza.texaspoker
+package io.github.nadzwyczajnaGrupaRobocza.texaspoker
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
