+++ androidApp/app/src/main/java/io/github/nadzwyczajnaGrupaRobocza/texaspoker/MainActivity.kt
@@ -1,4 +1,4 @@
-package io.github.NadzwyczajnaGrupaRobocza.texaspoker
+package io.github.nadzwyczajnaGrupaRobocza.texaspoker
 
 import android.os.Bundle
 import android.view.Menu
