+++ androidApp/app/src/main/java/io/github/nadzwyczajnaGrupaRobocza/texaspoker/game/Card.kt
@@ -1,0 +1,2 @@
+package io.github.nadzwyczajnaGrupaRobocza.texaspoker.game
+
