+++ androidApp/app/src/main/java/io/github/nadzwyczajnaGrupaRobocza/texaspoker/ui/gallery/GalleryViewModel.kt
@@ -1,13 +1,13 @@
-package io.github.NadzwyczajnaGrupaRobocza.texaspoker.ui.slideshow
+package io.github.nadzwyczajnaGrupaRobocza.texaspoker.ui.gallery
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class GalleryViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
