+++ androidApp/app/src/main/java/io/github/nadzwyczajnaGrupaRobocza/texaspoker/ui/slideshow/SlideshowFragment.kt
@@ -1,4 +1,4 @@
-package io.github.NadzwyczajnaGrupaRobocza.texaspoker.ui.slideshow
+package io.github.nadzwyczajnaGrupaRobocza.texaspoker.ui.slideshow
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,7 +8,7 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
-import io.github.NadzwyczajnaGrupaRobocza.texaspoker.R
+import io.github.nadzwyczajnaGrupaRobocza.texaspoker.R
 
 class SlideshowFragment : Fragment() {
