+++ androidApp/app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="io.github.NadzwyczajnaGrupaRobocza.texaspoker.ui.home.HomeFragment"
+        android:name="io.github.nadzwyczajnaGrupaRobocza.texaspoker.ui.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="io.github.NadzwyczajnaGrupaRobocza.texaspoker.ui.gallery.GalleryFragment"
+        android:name="io.github.nadzwyczajnaGrupaRobocza.texaspoker.ui.gallery.GalleryFragment"
         android:label="@string/menu_gallery"
         tools:layout="@layout/fragment_gallery" />
 
     <fragment
         android:id="@+id/nav_slideshow"
-        android:name="io.github.NadzwyczajnaGrupaRobocza.texaspoker.ui.slideshow.SlideshowFragment"
+        android:name="io.github.nadzwyczajnaGrupaRobocza.texaspoker.ui.slideshow.SlideshowFragment"
         android:label="@string/menu_slideshow"
         tools:layout="@layout/fragment_slideshow" />
 </navigation>
