+++ androidApp/app/src/test/java/io/github/nadzwyczajnaGrupaRobocza/texaspoker/game/CardTest.kt
@@ -1,0 +1,3 @@
+package io.github.nadzwyczajnaGrupaRobocza.texaspoker.game
+
+
